Add order + price type changes

+++ EPharm/EPharm.Domain/Dtos/OrderDto/ProductSummary.cs
@@ -3,6 +3,6 @@
 public class ProductSummary
 {
     public string Name { get; set; }
-    public int Value { get; set; }
+    public decimal Value { get; set; }
     public int Quantity { get; set; }
 }
